pyproject.toml
[project]
name = "geekpedia-util"
version = "0.0.0"
description = "Utility for working with Geekpedia"
readme = "README.md"
authors = [
    { name = "Justin Stout", email = "justin@jstout.us" }
]
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "lxml-html-clean>=0.4.2",
    "nanoid>=2.0.0",
    "newspaper3k>=0.2.8",
    "nltk>=3.9.1",
    "pyyaml>=6.0.2",
    "readability-lxml>=0.8.4.1",
    "typer>=0.16.0",
]

[project.scripts]
bede = "geekpedia_util.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "bump-my-version>=1.2.0",
    "invoke>=2.2.0",
    "pydocstyle>=6.3.0",
    "pydotenv>=0.0.7",
    "pytest>=8.4.1",
    "ruff>=0.12.0",
    "ty>=0.0.1a12",
]

[tool.ruff]
line-length = 110
target-version = "py313"
fix = true  # Automatically fix issues

[tool.ruff.lint]
select = ["E", "F", "I"]  # Enable specific rules (E: pycodestyle, F: Pyflakes, I: isort)

[tool.bumpversion]
current_version = "0.0.0"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
ignore_missing_files = false
tag = false
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = false
message = "Bump version: {current_version} → {new_version}"
moveable_tags = []
commit_args = ""
setup_hooks = []
pre_commit_hooks = []
post_commit_hooks = []

[[tool.bumpversion.files]]
filename = "pyproject.toml"

[[tool.bumpversion.files]]
filename = "src/geekpedia_util/__init__.py"
